+++ catalog.go
@@ -19,6 +19,8 @@ var (
 	FSMLBBattingBoxScore         Feed     = Feed(string(FS) + " mlb batting box score")
 	FSMLBPitchingBoxScore        Feed     = Feed(string(FS) + " mlb pitching box score")
 	FSMLBProbableStartingPitcher Feed     = Feed(string(FS) + " mlb pitching box score")
+	FSMLBOddsTotal               Feed     = Feed(string(FS) + " mlb odds total")
+	FSMLBOddsMoneyLine           Feed     = Feed(string(FS) + " mlb odds money line")
 	FSNFLMatchup                 Feed     = Feed(string(FS) + " nfl matchup")
 	FSNCAABMatchup               Feed     = Feed(string(FS) + " ncaab matchup")