+++ dataprovider/foxsports/jsonresponse/matchup_comparison_mlb.go
@@ -8,6 +8,22 @@ type MLBMatchupComparison struct {
 		HomePitcher MLBProbableStartingPitcher `json:"rightEntity"`
 		AwayPitcher MLBProbableStartingPitcher `json:"leftEntity"`
 	} `json:"featuredPairing"`
+
+	BetSection *struct {
+		Name string `json:"name"` // "name": "ODDS"
+		Bets []struct {
+			Template string `json:"template"` // "template": "market"
+			Model    struct {
+				Subtitle string `json:"subtitle"` // "subtitle": "RUN LINE" | // "subtitle": "TEAM TO WIN" | // "subtitle": "TOTAL"
+				MainText string `json:"mainText"` // "mainText": "The Rays must win by 2 runs or more to cover the run line"
+				Odds     []struct {
+					Text    string `json:"text"`    // "text": "-108"
+					SubText string `json:"subText"` // "subText": "CLE" | // "subText": "OVER 9"
+					Success *bool  `json:"success"` // "success": true
+				} `json:"odds"`
+			} `json:"model"`
+		} `json:"bets"`
+	} `json:"betSection"`
 }
 
 type MLBProbableStartingPitcher struct {